% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{aux.pkgstat}
\alias{aux.pkgstat}
\title{Show the number of functions for \pkg{Rdimtools}.}
\usage{
aux.pkgstat()
}
\description{
This function is mainly used for tracking progress for this package.
}
\examples{
## run with following command
aux.pkgstat()

}
