% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOBEWORKED_feature_kisung.R
\name{do.kisung}
\alias{do.kisung}
\title{Kisung's}
\usage{
do.kisung(X, ndim = 2, qthr = 0.75)
}
\description{
idea
1) compute distance correlation
2) quantile and threshold (Aij = 1 if DCorr(i,j) >= 'x' quantile)
3) shortest path
4) PAM -> partition id -> use as feature indices
}
\keyword{internal}
