% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_SDA.R
\name{do.sda}
\alias{do.sda}
\title{Semi-Supervised Discriminant Analysis}
\usage{
do.sda(X, label, ndim = 2, type = c("proportion", 0.1), alpha = 1,
  beta = 1)
}
\description{
Semi-Supervised Discriminant Analysis
}
\examples{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=33)-100
dt2  = aux.gensamples(n=33)
dt3  = aux.gensamples(n=33)+100

## merge the data and create a label correspondingly
Y      = rbind(dt1,dt2,dt3)
label  = c(rep(1,33), rep(2,33), rep(3,33))

#

}
\author{
Kisung You
}
