% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gdistnn.R
\name{est.gdistnn}
\alias{est.gdistnn}
\title{Intrinsic Dimension Estimation based on Manifold Assumption and Graph Distance}
\usage{
est.gdistnn(X, k = 5, k1 = 3, k2 = 10)
}
\description{
Intrinsic Dimension Estimation based on Manifold Assumption and Graph Distance
}
\examples{
\dontrun{
## create 3 datasets of intrinsic dimension 2.
X1 = aux.gensamples(dname="swiss")
X2 = aux.gensamples(dname="ribbon")
X3 = aux.gensamples(dname="saddle")

## acquire an estimate for intrinsic dimension
out1 = est.gdistnn(X1, k=10)
out2 = est.gdistnn(X2, k=10)
out3 = est.gdistnn(X3, k=10)

## print the results
sprintf("* est.gdistnn : estimated dimension for 'swiss'  data is \%.2f.",out1$estdim)
sprintf("* est.gdistnn : estimated dimension for 'ribbon' data is \%.2f.",out2$estdim)
sprintf("* est.gdistnn : estimated dimension for 'saddle' data is \%.2f.",out3$estdim)
}

}
\references{
\insertRef{he_intrinsic_2014}{Rdimtools}
}
\author{
Kisung You
}
