% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mindkl.R
\name{est.mindkl}
\alias{est.mindkl}
\title{MiNDkl}
\usage{
est.mindkl(X, k = 5)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations.}

\item{k}{the neighborhood size for defining locality.}
}
\value{
a named list containing containing \describe{
\item{estdim}{the global estimated dimension.}
}
}
\description{
It is a minimum neighbor distance estimator of the intrinsic dimension based on Kullback Leibler divergence estimator.
}
\examples{
## create 3 datasets of intrinsic dimension 2.
X1 = aux.gensamples(dname="swiss")
X2 = aux.gensamples(dname="ribbon")
X3 = aux.gensamples(dname="saddle")

## acquire an estimate for intrinsic dimension
out1 = est.mindkl(X1, k=5)
out2 = est.mindkl(X2, k=5)
out3 = est.mindkl(X3, k=5)

## print the results
line1 = paste0("* est.mindkl : 'swiss'  estiamte is ",round(out1$estdim,2))
line2 = paste0("* est.mindkl : 'ribbon' estiamte is ",round(out2$estdim,2))
line3 = paste0("* est.mindkl : 'saddle' estiamte is ",round(out3$estdim,2))
cat(paste0(line1,"\n",line2,"\n",line3))

}
\references{
\insertRef{lombardi_minimum_2011}{Rdimtools}
}
\seealso{
\code{\link{est.mindml}}
}
\author{
Kisung You
}
