% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_CNPE.R
\name{do.cnpe}
\alias{do.cnpe}
\title{Complete Neighborhood Preserving Embedding}
\usage{
do.cnpe(X, ndim = 2, type = c("proportion", 0.1), preprocess = c("center",
  "decorrelate", "whiten"))
}
\description{
Complete Neighborhood Preserving Embedding
}
\author{
Kisung You
}
