% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_LEA.R
\name{do.lea}
\alias{do.lea}
\title{Locally Linear Embedded Eigenspace Analysis}
\usage{
do.lea(X, ndim = 2, type = c("proportion", 0.1), symmetric = c("union",
  "intersect", "asymmetric"), preprocess = c("center", "decorrelate",
  "whiten"))
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{type}{a vector of neighborhood graph construction. Following types are supported;
\code{c("knn",k)}, \code{c("enn",radius)}, and \code{c("proportion",ratio)}.
Default is \code{c("proportion",0.1)}, connecting about 1/10 of nearest data points
among all data points. See also \code{\link{aux.graphnbd}} for more details.}

\item{symmetric}{one of \code{"intersect"}, \code{"union"} or \code{"asymmetric"} is supported. Default is \code{"union"}. See also \code{\link{aux.graphnbd}} for more details.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "center" and other options of "decorrelate" and "whiten"
are supported. See also \code{\link{aux.preprocess}} for more details.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
\item{projection}{a \eqn{(p\times ndim)} whose columns are basis for projection.}
}
}
\description{
Locally Linear Embedding (LLE) is a powerful nonlinear manifold learning method. This method,
Locally Linear Embedded Eigenspace Analysis - LEA, in short - is a linear approximation to LLE,
similar to Neighborhood Preserving Embedding. In our implementation, the choice of weight binarization
is removed in order to respect original work. For 1-dimensional projection, which is rarely performed,
authors provided a detour for rank correcting mechanism but it is omitted for practical reason.
}
\examples{
\dontrun{
## generate samples
X <- aux.gensamples(n=123)

## compare LEA with LLE and another approximation NPE
out1 <- do.lle(X, ndim=2)
out2 <- do.npe(X, ndim=2)
out3 <- do.lea(X, ndim=2)

## visual comparison
par(mfrow=c(1,3))
plot(out1$Y[,1], out1$Y[,2], main="LLE")
plot(out2$Y[,1], out2$Y[,2], main="NPE")
plot(out3$Y[,1], out3$Y[,2], main="LEA")
}

}
\references{
\insertRef{fu_locally_2005}{Rdimtools}
}
\seealso{
\code{\link{do.npe}}
}
\author{
Kisung You
}
