% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_LFDA.R
\name{do.lfda}
\alias{do.lfda}
\title{Local Fisher Discriminant Analysis}
\usage{
do.lfda(X, label, ndim = 2, preprocess = c("center", "decorrelate",
  "whiten"), type = c("proportion", 0.1), symmetric = c("union",
  "intersect", "asymmetric"), localscaling = TRUE)
}
\description{
Local Fisher Discriminant Analysis
}
\references{
\insertRef{sugiyama_local_2006}{Rdimtools}

\insertRef{zelnik-manor_self-tuning_2005}{Rdimtools}
}
\author{
Kisung You
}
