% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_LPFDA.R
\name{do.lpfda}
\alias{do.lpfda}
\title{Locality Preserving Fisher Discriminant Analysis}
\usage{
do.lpfda(X, label, ndim = 2, type = c("proportion", 0.1),
  preprocess = c("center", "whiten", "decorrelate"), t = 10)
}
\description{
Locality Preserving Fisher Discriminant Analysis
}
\examples{
\dontrun{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=33)-100
dt2  = aux.gensamples(n=33)
dt3  = aux.gensamples(n=33)+100

## merge the data and create a label correspondingly
X      = rbind(dt1,dt2,dt3)
label  = c(rep(1,33), rep(2,33), rep(3,33))

## try different kernel bandwidth parameters
out1 = do.lpfda(X, label, t=0.01)
out2 = do.lpfda(X, label, t=1)
out3 = do.lpfda(X, label, t=100)

## visualize
par(mfrow=c(1,3))
plot(out1$Y[,1], out1$Y[,2], main="MMSD::C=0.01")
plot(out2$Y[,1], out2$Y[,2], main="MMSD::C=1")
plot(out3$Y[,1], out3$Y[,2], main="MMSD::C=100")
}

}
\references{
\insertRef{zhao_locality_2009}{Rdimtools}
}
\author{
Kisung You
}
