% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_MCFS.R
\name{do.mcfs}
\alias{do.mcfs}
\title{Multi-Cluster Feature Selection}
\usage{
do.mcfs(X, ndim = 2, type = c("proportion", 0.1), preprocess = c("null",
  "center", "scale", "cscale", "whiten", "decorrelate"),
  K = max(round(nrow(X)/5), 2), lambda = 1, t = 10)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{type}{a vector of neighborhood graph construction. Following types are supported;
\code{c("knn",k)}, \code{c("enn",radius)}, and \code{c("proportion",ratio)}.
Default is \code{c("proportion",0.1)}, connecting about 1/10 of nearest data points
among all data points. See also \code{\link{aux.graphnbd}} for more details.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "null". See also \code{\link{aux.preprocess}} for more details.}

\item{K}{assumed number of clusters in the original dataset.}

\item{lambda}{\eqn{\ell_1} regularization parameter in \eqn{(0,\infty)}.}

\item{t}{bandwidth parameter for heat kernel in \eqn{(0,\infty)}.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{featidx}{a length-\eqn{ndim} vector of indices with highest scores.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
\item{projection}{a \eqn{(p\times ndim)} whose columns are basis for projection.}
}
}
\description{
Multi-Cluster Feature Selection (MCFS) is an unsupervised feature selection method. Based on
a multi-cluster assumption, it aims at finding meaningful features using sparse reconstruction of
spectral basis using LASSO.
}
\examples{
\dontrun{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=33)-100
dt2  = aux.gensamples(n=33)
dt3  = aux.gensamples(n=33)+100

## merge the data and create a label correspondingly
X      = rbind(dt1,dt2,dt3)
label  = c(rep(1,33), rep(2,33), rep(3,33))

## try different regularization parameters
out1 = do.mcfs(X, lambda=0.01)
out2 = do.mcfs(X, lambda=0.1)
out3 = do.mcfs(X, lambda=1)

## visualize
par(mfrow=c(1,3))
plot(out1$Y[,1], out1$Y[,2], main="lambda=0.01")
plot(out2$Y[,1], out2$Y[,2], main="lambda=0.1")
plot(out3$Y[,1], out3$Y[,2], main="lambda=1")
}

}
\references{
\insertRef{cai_unsupervised_2010}{Rdimtools}
}
\author{
Kisung You
}
