% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_MMP.R
\name{do.mmp}
\alias{do.mmp}
\title{Maximum Margin Projection}
\usage{
do.mmp(X, label, ndim = 2, preprocess = c("center", "decorrelate",
  "whiten"), numk = max(ceiling(nrow(X)/10), 2), alpha = 0.5, gamma = 50)
}
\description{
Maximum Margin Projection
}
\examples{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=33)-100
dt2  = aux.gensamples(n=33)
dt3  = aux.gensamples(n=33)+100

## merge the data and create a label correspondingly
X      = rbind(dt1,dt2,dt3)
label  = c(rep(1,33), rep(2,33), rep(3,33))

## copy a label and let 20\% of elements be missing
nlabel = length(label)
nmissing = round(nlabel*0.20)
label_missing = label
label_missing[sample(1:nlabel, nmissing)]=NA

## compare with PCA case for full-label case
## for missing label case from MMP computation
out1 = do.pca(X, ndim=2)
out2 = do.mmp(X, label_missing, numk=10)

## visualize
par(mfrow=c(1,2))
plot(out1$Y[,1], out1$Y[,2], main="PCA projection")
plot(out2$Y[,1], out2$Y[,2], main="20\% missing labels")

}
\author{
Kisung You
}
