% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_NOLPP.R
\name{do.nolpp}
\alias{do.nolpp}
\title{Nonnegative Orthogonal Locality Preserving Projection}
\usage{
do.nolpp(X, ndim = 2, type = c("proportion", 0.1),
  preprocess = c("center", "decorrelate", "whiten"), t = 1,
  maxiter = 1000, reltol = 1e-05)
}
\description{
Nonnegative Orthogonal Locality Preserving Projection
}
\references{
\insertRef{zafeiriou_nonnegative_2010}{Rdimtools}
}
\author{
Kisung You
}
