% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_NONPP.R
\name{do.nonpp}
\alias{do.nonpp}
\title{Nonnegative Orthogonal Neighborhood Preserving Projections}
\usage{
do.nonpp(X, ndim = 2, type = c("proportion", 0.1),
  preprocess = c("center", "decorrelate", "whiten"), maxiter = 1000,
  reltol = 1e-05)
}
\description{
Nonnegative Orthogonal Neighborhood Preserving Projections
}
\references{
\insertRef{zafeiriou_nonnegative_2010}{Rdimtools}
}
\author{
Kisung You
}
