% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_NPCA.R
\name{do.npca}
\alias{do.npca}
\title{Nonnegative Principal Component Analysis}
\usage{
do.npca(X, ndim = 2, preprocess = c("center", "decorrelate", "whiten"),
  maxiter = 1000, reltol = 1e-05)
}
\description{
Nonnegative Principal Component Analysis
}
\references{
\insertRef{zafeiriou_nonnegative_2010}{Rdimtools}
}
\author{
Kisung You
}
