% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_OLDA.R
\name{do.olda}
\alias{do.olda}
\title{Orthogonal Linear Discriminant Analysis}
\usage{
do.olda(X, label, ndim = 2, preprocess = c("center", "whiten",
  "decorrelate"))
}
\description{
Orthogonal Linear Discriminant Analysis
}
\author{
Kisung You
}
