% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_RPCAG.R
\name{do.rpcag}
\alias{do.rpcag}
\title{Robust Principal Component Analysis via Geometric Median}
\usage{
do.rpcag(
  X,
  ndim = 2,
  k = 5,
  preprocess = c("center", "scale", "cscale", "whiten", "decorrelate")
)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{k}{the number of subsets for \code{X} to be divided.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "center". See also \code{\link{aux.preprocess}} for more details.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
\item{projection}{a \eqn{(p\times ndim)} whose columns are basis for projection.}
}
}
\description{
This function robustifies the traditional PCA via an idea of geometric median.
To describe, the given data is first split into \code{k} subsets for each sample
covariance is attained. According to the paper, the median covariance is computed
under Frobenius norm and projection is extracted from the largest eigenvectors.
}
\examples{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=50)-100
dt2  = aux.gensamples(n=50)
dt3  = aux.gensamples(n=50)+100

## merge the data and create a label correspondingly
X      = rbind(dt1,dt2,dt3)
label  = rep(1:3, each=50)

## try different numbers for subsets
out1 = do.rpcag(X, ndim=2, k=2)
out2 = do.rpcag(X, ndim=2, k=5)
out3 = do.rpcag(X, ndim=2, k=10)

## visualize
opar <- par(mfrow=c(1,3), no.readonly=TRUE)
plot(out1$Y[,1], out1$Y[,2], col=label, main="k=2")
plot(out2$Y[,1], out2$Y[,2], col=label, main="k=5")
plot(out3$Y[,1], out3$Y[,2], col=label, main="k=10")
par(opar)

}
\references{
\insertRef{minsker_geometric_2015}{Rdimtools}
}
\author{
Kisung You
}
