% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_RSIR.R
\name{do.rsir}
\alias{do.rsir}
\title{Regularized Sliced Inverse Regression}
\usage{
do.rsir(X, response, ndim = 2, h = max(2, round(nrow(X)/5)),
  preprocess = c("center", "decorrelate", "whiten"), regmethod = c("Ridge",
  "Tikhonov", "PCA", "PCARidge", "PCATikhonov"), tau = 1, numpc = ndim)
}
\description{
Regularized Sliced Inverse Regression
}
\examples{
## generate swiss roll with auxiliary dimensions
## it follows reference example from LSIR paper.
n = 123
theta = runif(n)
h     = runif(n)
t     = (1+2*theta)*(3*pi/2)
X     = array(0,c(n,10))
X[,1] = t*cos(t)
X[,2] = 21*h
X[,3] = t*sin(t)
X[,4:10] = matrix(runif(7*n), nrow=n)

## corresponding response vector
y = sin(5*pi*theta)+(runif(n)*sqrt(0.1))

## try with different regularization methods
out1 = do.rsir(X, y, regmethod="Ridge")
out2 = do.rsir(X, y, regmethod="Tikhonov")
out3 = do.rsir(X, y, regmethod="PCA")
out4 = do.rsir(X, y, regmethod="PCARidge")
out5 = do.rsir(X, y, regmethod="PCATikhonov")
outsir = do.sir(X, y)

## visualize
par(mfrow=c(1,3))
plot(out1$Y[,1], out1$Y[,2], main="2 slices")
plot(out2$Y[,1], out2$Y[,2], main="10 slices")
plot(out3$Y[,1], out3$Y[,2], main="25 slices")

}
\seealso{
\code{\link{do.sir}}
}
\author{
Kisung You
}
