% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_SDLPP.R
\name{do.sdlpp}
\alias{do.sdlpp}
\title{Sample-Dependent Locality Preserving Projection}
\usage{
do.sdlpp(X, ndim = 2, t = 1, preprocess = c("center", "decorrelate",
  "whiten"))
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations.}

\item{ndim}{an integer-valued target dimension.}

\item{t}{kernel bandwidth in \eqn{(0,\infty)}.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "center" and other options of "decorrelate" and "whiten"
are supported. See also \code{\link{aux.preprocess}} for more details.}
}
\description{
Sample-Dependent Locality Preserving Projection
}
\examples{
## generate data
X <-


}
\references{
\insertRef{yang_sample-dependent_2010}{Rdimtools}
}
\author{
Kisung You
}
