% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_SLPE.R
\name{do.slpe}
\alias{do.slpe}
\title{Supervised Locality Pursuit Embedding}
\usage{
do.slpe(X, label, ndim = 2, preprocess = c("center", "decorrelate",
  "whiten"))
}
\description{
Supervised Locality Pursuit Embedding
}
\references{
\insertRef{zheng_supervised_2006}{Rdimtools}
}
\seealso{
\code{\link{do.lpe}}
}
\author{
Kisung You
}
