% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ULDA.R
\name{do.ulda}
\alias{do.ulda}
\title{Uncorrelated Linear Discriminant Analysis}
\usage{
do.ulda(X, label, ndim = 2, preprocess = c("center", "whiten",
  "decorrelate"))
}
\description{
Uncorrelated Linear Discriminant Analysis
}
\examples{
\dontrun{
## generate data of 3 types with clear difference
dt1  = aux.gensamples(n=33)-100
dt2  = aux.gensamples(n=33)
dt3  = aux.gensamples(n=33)+100

## merge the data and create a label correspondingly
X      = rbind(dt1,dt2,dt3)
label  = c(rep(1,33), rep(2,33), rep(3,33))

## compare with LDA
out1 = do.lda(X, label)
out2 = do.ulda(X, label)

## visualize
par(mfrow=c(1,2))
plot(out1$Y[,1], out1$Y[,2])
plot(out2$Y[,1], out2$Y[,2])
}

}
