% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_CISOMAP.R
\name{do.cisomap}
\alias{do.cisomap}
\title{Conformal Isometric Feature Mapping}
\usage{
do.cisomap(X, ndim = 2, type = c("proportion", 0.1), symmetric = "union",
  weight = TRUE, preprocess = "center")
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{type}{a vector of neighborhood graph construction. Following types are supported;
\code{c("knn",k)}, \code{c("enn",radius)}, and \code{c("proportion",ratio)}.
Default is \code{c("proportion",0.1)}, connecting about 1/10 of nearest data points
among all data points. See also \code{\link{aux.graphnbd}} for more details.}

\item{symmetric}{one of \code{"intersect"}, \code{"union"} or \code{"asymmetric"} is supported. Default is \code{"union"}. See also \code{\link{aux.graphnbd}} for more details.}

\item{weight}{\code{TRUE} to perform Isomap on weighted graph, or \code{FALSE} otherwise.}

\item{preprocess}{an additional option for preprocessing the data.
Default is ``center'' and other methods of ``decorrelate'', or ``whiten''
are supported. See also \code{\link{aux.preprocess}} for more details.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
}
}
\description{
Conformal Isomap(C-Isomap) is a variant of a celebrated method of Isomap. It aims at, rather than
preserving full isometry, maintaining infinitestimal angles - conformality - in that
it alters geodesic distance to reflect scale information.
}
\examples{
\dontrun{
## generate data
X <- aux.gensamples(dname="cswiss",n=500)

## 1. original Isomap
output1 <- do.isomap(X,ndim=2)

## 2. C-Isomap
output2 <- do.cisomap(X,ndim=2)

## 3. C-Isomap on a binarized graph
output3 <- do.cisomap(X,ndim=2,weight=FALSE)

## Visualize three different projections
par(mfrow=c(1,3))
plot(output1$Y[,1],output1$Y[,2],main="Isomap")
plot(output2$Y[,1],output2$Y[,2],main="C-Isomap")
plot(output3$Y[,1],output3$Y[,2],main="Binarized C-Isomap")
}

}
\references{
\insertRef{silva_global_2003}{Rdimtools}
}
\author{
Kisung You
}
