% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdimtools-package.R
\docType{package}
\name{package-Rdimtools}
\alias{package-Rdimtools}
\alias{Rdimtools-package}
\title{Dimension Reduction and Estimation Methods}
\description{
\pkg{Rdimtools} is an R suite of a number of dimension reduction and estimation methods
implemented using \pkg{RcppArmadillo} for efficient computations. Please see the reference
from the \href{http://kyoustat.com/Rdimtools}{package webpage}.
}
