% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_MIFS.R
\name{do.mifs}
\alias{do.mifs}
\title{Mutual Information for Selecting Features}
\usage{
do.mifs(
  X,
  label,
  ndim = 2,
  beta = 0.75,
  discretize = c("default", "histogram"),
  preprocess = c("null", "center", "scale", "cscale", "whiten", "decorrelate")
)
}
\description{
Mutual Information for Selecting Features
}
\examples{
\donttest{
## use iris data
## it is known that feature 3 and 4 are more important.
data(iris)
iris.dat = as.matrix(iris[,1:4])
iris.lab = as.factor(iris[,5])

## try different beta values
out1 = do.mifs(iris.dat, iris.lab, discretize="histogram", beta=0.5)
out2 = do.mifs(iris.dat, iris.lab, discretize="histogram", beta=0.75)
out3 = do.mifs(iris.dat, iris.lab, discretize="histogram", beta=1)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(2,2))
plot(out1$Y, col=iris.lab, main="ratio")
plot(out2$Y, col=iris.lab, main="diff/lambda=0")
plot(out3$Y, col=iris.lab, main="diff/lambda=0.5")
par(opar)
}

}
\author{
Kisung You
}
\concept{feature_methods}
