% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ADR.R
\name{do.adr}
\alias{do.adr}
\title{Adaptive Dimension Reduction}
\usage{
do.adr(
  X,
  ndim = 2,
  preprocess = c("center", "scale", "cscale", "decorrelate", "whiten"),
  maxiter = 10,
  abstol = 0.001
)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations.}

\item{ndim}{an integer-valued target dimension.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "center". See also \code{\link{aux.preprocess}} for more details.}

\item{maxiter}{maximum number of iterations allowed.}

\item{abstol}{stopping criterion for incremental change in projection matrix.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
\item{projection}{a \eqn{(p\times ndim)} whose columns are basis for projection.}
}
}
\description{
Adaptive Dimension Reduction (ADR) iteratively finds the best subspace to perform data clustering. It can be regarded as
one of remedies for clustering in high dimensional space. Eigenvectors of a between-cluster scatter matrix are used
as basis of projection.
}
\examples{
## load iris data
data(iris)
X     = as.matrix(iris[,1:4])
label = as.factor(iris$Species)

## try different tolerance level
out1 = do.adr(X, abstol=1e-2)
out2 = do.adr(X, abstol=1e-3)
out3 = do.adr(X, abstol=1e-4)

## visualize
opar <- par(mfrow=c(1,3), no.readonly=TRUE)
plot(out1$Y, main="tol=1e-2", col=label, pch=19, cex=0.5)
plot(out2$Y, main="tol=1e-3", col=label, pch=19, cex=0.5)
plot(out3$Y, main="tol=1e-4", col=label, pch=19, cex=0.5)
par(opar)

}
\references{
\insertRef{ding_adaptive_2002}{Rdimtools}
}
\seealso{
\code{\link{do.ldakm}}
}
\author{
Kisung You
}
