% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_EXTLPP.R
\name{do.extlpp}
\alias{do.extlpp}
\title{Extended Locality Preserving Projection}
\usage{
do.extlpp(X, ndim = 2, numk = max(ceiling(nrow(X)/10), 2),
  preprocess = c("center", "decorrelate", "whiten"))
}
\description{
Extended Locality Preserving Projection
}
