% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_LPE.R
\name{do.lpe}
\alias{do.lpe}
\title{Locality Pursuit Embedding}
\usage{
do.lpe(X, ndim = 2, preprocess = c("center", "decorrelate", "whiten"),
  numk = max(ceiling(nrow(X)/10), 2))
}
\description{
Locality Pursuit Embedding
}
\examples{
\dontrun{
## generate swiss roll with auxiliary dimensions
n = 100
theta = runif(n)
h     = runif(n)
t     = (1+2*theta)*(3*pi/2)
X     = array(0,c(n,10))
X[,1] = t*cos(t)
X[,2] = 21*h
X[,3] = t*sin(t)
X[,4:10] = matrix(runif(7*n), nrow=n)

## try with different neighborhood sizes
out1 = do.lpe(X, numk=5)
out2 = do.lpe(X, numk=10)
out3 = do.lpe(X, numk=25)

## visualize
par(mfrow=c(1,3))
plot(out1$Y[,1], out1$Y[,2], main="LPE::numk=5")
plot(out2$Y[,1], out2$Y[,2], main="LPE::numk=10")
plot(out3$Y[,1], out3$Y[,2], main="LPE::numk=25")
}

}
\references{
\insertRef{min_locality_2004}{Rdimtools}
}
\author{
Kisung You
}
