% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_LSIR.R
\name{do.lsir}
\alias{do.lsir}
\title{Localized Sliced Inverse Regression}
\usage{
do.lsir(X, response, ndim = 2, h = max(2, round(nrow(X)/5)),
  preprocess = c("center", "decorrelate", "whiten"), numk = max(2,
  round(nrow(X)/10)), tau = 1)
}
\description{
Localized Sliced Inverse Regression
}
\seealso{
\code{\link{do.sir}}
}
\author{
Kisung You
}
