% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_NPE.R
\name{do.npe}
\alias{do.npe}
\title{Neighborhood Preserving Embedding}
\usage{
do.npe(X, ndim = 2, type = c("proportion", 0.1), symmetric = "union",
  weight = TRUE, preprocess = "null", regtype = FALSE, regparam = 1)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{type}{a vector of neighborhood graph construction. Following types are supported;
\code{c("knn",k)}, \code{c("enn",radius)}, and \code{c("proportion",ratio)}.
Default is \code{c("proportion",0.1)}, connecting about 1/10 of nearest data points
among all data points. See also \code{\link{aux.graphnbd}} for more details.}

\item{symmetric}{one of \code{"intersect"}, \code{"union"} or \code{"asymmetric"} is supported. Default is \code{"union"}. See also \code{\link{aux.graphnbd}} for more details.}

\item{weight}{\code{TRUE} to perform NPE on weighted graph, or \code{FALSE} otherwise.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "null" and three options of "center", "decorrelate", or "whiten"
are supported. See also \code{\link{aux.preprocess}} for more details.}

\item{regtype}{\code{FALSE} for not applying automatic Tikhonov Regularization,
or \code{TRUE} otherwise.}

\item{regparam}{a positive real number for Regularization. Default value is 1.}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{eigval}{a vector of eigenvalues corresponding to basis expansion in an ascending order.}
\item{projection}{a \eqn{(p\times ndim)} whose columns are basis for projection.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
}
}
\description{
\code{do.npe} performs a linear dimensionality reduction using Neighborhood Preserving
Embedding (NPE) proposed by He et al (2005). It can be regarded as a linear approximation
to Laplacian Eigenmaps.
}
\details{
Like Locally Linear Embedding (LLE), it is possible for the weight matrix being rank deficient.
If \code{regtype} is set to \code{TRUE} with a proper value of \code{regparam}, it will
perform Tikhonov regularization as designated. When regularization is needed
with \code{regtype} parameter to be \code{FALSE}, it will automatically find a suitable
regularization parameter and put penalty for stable computation. See also
\code{\link{do.lle}} for more details.
}
\examples{
# generate data
X <- aux.gensamples(n=333)

## 1. connecting 5\% of data for graph construction.
output1 <- do.npe(X,ndim=2,type=c("proportion",0.05))

## 2. constructing 25\%-connected graph with regularization parameter
output2 <- do.npe(X,ndim=2,type=c("proportion",0.2),symmetric='intersect',regparam=1.0)

## 3. constructing half-connected graph with reg parameter = 10.0.
output3 <- do.npe(X,ndim=2,type=c("proportion",0.5),regparam=10.0)

## Visualize three different projections
par(mfrow=c(1,3))
plot(output1$Y[,1],output1$Y[,2],main="5\%")
plot(output2$Y[,1],output2$Y[,2],main="25\%")
plot(output3$Y[,1],output3$Y[,2],main="50\%")

}
\references{
\insertRef{he_neighborhood_2005}{Rdimtools}
}
\author{
Kisung You
}
