% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_ONPP.R
\name{do.onpp}
\alias{do.onpp}
\title{Orthogonal Neighborhood Preserving Projections}
\usage{
do.onpp(X, ndim = 2, type = c("proportion", 0.1), preprocess = c("center",
  "decorrelate", "whiten"))
}
\description{
Orthogonal Neighborhood Preserving Projections
}
\author{
Kisung You
}
