% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_SAMMON.R
\name{do.sammon}
\alias{do.sammon}
\title{Sammon Mapping}
\usage{
do.sammon(
  X,
  ndim = 2,
  preprocess = c("null", "center", "scale", "cscale", "decorrelate", "whiten"),
  initialize = c("pca", "random")
)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix or data frame whose rows are observations
and columns represent independent variables.}

\item{ndim}{an integer-valued target dimension.}

\item{preprocess}{an additional option for preprocessing the data.
Default is "null". See also \code{\link{aux.preprocess}} for more details.}

\item{initialize}{\code{"random"} or \code{"pca"}; the former performs
fast random projection (see also \code{\link{do.rndproj}}) and the latter
performs standard PCA (see also \code{\link{do.pca}}).}
}
\value{
a named list containing
\describe{
\item{Y}{an \eqn{(n\times ndim)} matrix whose rows are embedded observations.}
\item{trfinfo}{a list containing information for out-of-sample prediction.}
}
}
\description{
\code{do.sammon} is an implementation for Sammon mapping, one of the earliest
dimension reduction techniques that aims to find low-dimensional embedding
that preserves pairwise distance structure in high-dimensional data space.
}
\examples{
## load iris data
data(iris)
X     = as.matrix(iris[,1:4])
label = as.factor(iris$Species)

## compare two initialization
out1 = do.sammon(X,ndim=2)                   # random projection
out2 = do.sammon(X,ndim=2,initialize="pca")  # pca as initialization

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2))
plot(out1$Y, pch=19, col=label, main="out1:rndproj")
plot(out2$Y, pch=19, col=label, main="out2:pca")
par(opar)

}
\references{
Sammon, J.W. (1969) \emph{A Nonlinear Mapping for Data Structure Analysis}.
IEEE Transactions on Computers, C-18 5:401-409.

\insertRef{sammon_nonlinear_1969}{Rdimtools}
}
\author{
Kisung You
}
\concept{nonlinear_methods}
